Move interactions module under the user_interface subpackage
