+++ LTUAssistantPlus/skills/add_calendar_event_skill.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 
 import calendardb
-import interactions
+import user_interface.interactions
 import user_interface.speaking
 
 from nlp.universal_dependencies import ParsedUniversalDependencies
@@ -24,9 +24,9 @@ class AddCalendarEventSkill(Skill):
         verb_object = skill_input.noun
         event_str = verb_object
         if event_str == 'event':
-            event_sentence = interactions.ask_question('Okay, what is the event called?', skill_input.verbose)
-            day_sentence = interactions.ask_question('What day will this be on?', skill_input.verbose)
-            time_sentence = interactions.ask_question('What time will this start at?', skill_input.verbose)
+            event_sentence = user_interface.interactions.ask_question('Okay, what is the event called?', skill_input.verbose)
+            day_sentence = user_interface.interactions.ask_question('What day will this be on?', skill_input.verbose)
+            time_sentence = user_interface.interactions.ask_question('What time will this start at?', skill_input.verbose)
             cal_event = calendardb.CalendarEvent(event_sentence, day_sentence, time_sentence, '')
             calendardb.add_event(cal_event)
             feedback_sentence = 'Alright, I\'m putting down ' + str(cal_event) + '.'
