+++ LTUAssistantPlus/skills/send_email_skill.py
@@ -1,6 +1,5 @@
 #!/usr/bin/python3
 
-import interactions
 import user_interface.speaking
 import webbrowser
